% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiderplots.R
\name{funcCircleCoords}
\alias{funcCircleCoords}
\title{Spider Plot Position Grid lines}
\usage{
funcCircleCoords(center = c(0, 0), r = 1, npoints = 100)
}
\value{

}
\description{
Spider Plot Position Grid lines
}
\examples{

}
