% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{make_cluster_phylo}
\alias{make_cluster_phylo}
\title{RUn Cluster Plots}
\usage{
make_cluster_phylo(
  data,
  .id_col,
  .attributes = tidyselect::everything(),
  palette_col = "Dark2",
  body_bg_col = "white",
  min_num_clusts = 4,
  max_num_clusts = 6
)
}
\arguments{
\item{data}{data.frame for cluster plot}

\item{.id_col}{id column}

\item{.attributes}{selector for attributes}

\item{palette_col}{color palette for clusters}

\item{body_bg_col}{background color}

\item{min_num_clusts}{minimum number of clusters}

\item{max_num_clusts}{maximum number of clusters}
}
\value{
list of cluster plots
}
\description{
RUn Cluster Plots
}
