% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiderplots.R
\name{make_radar_plot}
\alias{make_radar_plot}
\title{Spider Plot}
\usage{
make_radar_plot(
  radar_plot_data,
  font_name,
  font_size,
  background_fill,
  values.radar = c(0, 75, 150)
)
}
\arguments{
\item{font_name}{}

\item{background_fill}{}
}
\value{

}
\description{
Spider Plot
}
\examples{

}
