% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sumtable.R
\name{make_sumtable}
\alias{make_sumtable}
\title{Sumtable}
\usage{
make_sumtable(
  data,
  .products_colname,
  .attributes_colname,
  .attribute_value_colname,
  .id_cols,
  selected_products,
  selected_attributes
)
}
\arguments{
\item{.products_colname}{products column}

\item{.attributes_colname}{attribute names column}

\item{.attribute_value_colname}{attribute values column}

\item{.id_cols}{id columns}

\item{selected_products}{vector of product names to be selected}

\item{selected_attributes}{vector of attributes names to be selected}

\item{dataio}{input dataframe}
}
\value{
c("tbl_df", "tbl", "data.frame")
}
\description{
Sumtable
}
\examples{
data <- mtcars
data$product <- rownames(data)

data <- data \%>\%
  tidyr::pivot_longer(cols = -c(product, cyl))

data_transformed <- make_sumtable(data = data,
                                  .products_colname = product,
                                  .attributes_colname = name,
                                  .attribute_value_colname = value,
                                  .id_cols = -value,
                                  selected_products = c("Merc 280", "Fiat 128", "Mazda RX4"),
                                  selected_attributes = c("mpg", "disp"))

}
